*) changes from Yemla:
Battlemaster: ML6 Throw Weapon- trying to find a solution to it, because it keeps saying that its not implemented i added in a fix for disarm but still no luck on the implement thing (is it a SVN issue?)

Convoker: PowerRend not to intrupt (because it handles spells that shouldnt inrupt) Speedwarp and Prescience Node Spell (the ones the nodes cast) changed from level 1 to 50 so resistchance is dramaticly changed.

Perfecter: ML8 to target RvR sickness as well as PvE.

Warlord: ML8 fixed to dmg the IF it heals someone around them.

Ablative: Fixed it to let the absorbtion reach 100% was 99% (this minor change is related to Database changes inc. from me in later donations xP)

AFHits: Fixed to increase both Hits and Armor...think it only increased hits.

Buffshear: Fixed it to randomly shear 1 of 30 spell types it does each proc instead of go thru the list shearing the target if they have 1 of the 30 in a row...example random number is 10 (str buff) if target doesnt have it, it isnt sheared, unlike the other way where if they didnt have strbuff it wood shear another buff because thats not livelike it randomly TRYS to shear 1 of the spell types. x)